자동 구정 정보 파인 분리
