+++ README.md
@@ -76,3 +76,4 @@
           - DispatcherServlet
     4. 인프라 빈 구성 정보의 분리
     5. 동적인 자동 구성 정보 등록
+    6. 자동 구성 정보 파일 분리
