+++ src/main/java/study/spring/config/EnableMyAutoConfiguration.java
@@ -1,11 +1,6 @@
-package study.spring.springboot.annotation;
+package study.spring.config;
 
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-import study.spring.springboot.config.DispatcherServletConfig;
-import study.spring.springboot.config.MyAutoConfigImportSelector;
-import study.spring.springboot.config.TomcatWebServerConfig;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
