+++ src/main/java/study/spring/config/MyAutoConfiguration.java
@@ -1,6 +1,5 @@
-package study.spring.springboot.annotation;
+package study.spring.config;
 
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 import java.lang.annotation.ElementType;
@@ -11,7 +10,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
 @Configuration
-@ComponentScan
-@EnableMyAutoConfiguration
-public @interface MySpringBootApplication {
+public @interface MyAutoConfiguration {
 }
