+++ src/main/java/study/spring/config/autoconfig/TomcatWebServerConfig.java
@@ -1,4 +1,4 @@
-package study.spring.springboot.config;
+package study.spring.config.autoconfig;
 
 import org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory;
 import org.springframework.boot.web.servlet.server.ServletWebServerFactory;
