+++ src/main/java/study/spring/springboot/HelloController.java
@@ -1,4 +1,4 @@
-package study.spring.springboot.core;
+package study.spring.springboot;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
