+++ src/main/java/study/spring/springboot/SimpleHelloService.java
@@ -1,4 +1,4 @@
-package study.spring.springboot.core;
+package study.spring.springboot;
 
 import org.springframework.stereotype.Service;
