--- src/main/java/study/spring/springboot/annotation/MyComponent.java
@@ -1,14 +1,0 @@
-package study.spring.springboot.annotation;
-
-import org.springframework.stereotype.Component;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-@Component
-public @interface MyComponent {
-}
