+++ src/main/resources/META-INF/spring/study.spring.config.MyAutoConfiguration.imports
@@ -1,0 +1,2 @@
+study.spring.config.autoconfig.DispatcherServletConfig
+study.spring.config.autoconfig.TomcatWebServerConfig
